add source

+++ services/i_map_service.py
@@ -3,5 +3,5 @@ from abc import ABC, abstractmethod
 
 class IMapService(ABC):
     @abstractmethod
-    def get_map(self, longitude: float, latitude: float, zoom: float):
+    def get_map(self, longitude: float, latitude: float, zoom: int):
         pass
